brains/Math2matics2.brain/outputs/code-creation/agent_1766547586805_j8aytnq/requirements.txt
numpy==1.26.4
matplotlib==3.8.4
Pillow==10.4.0
pytest==8.3.3
